+++ terminusdb_client/woqlquery/woql_query.py
@@ -1900,7 +1900,6 @@ class WOQLQuery:
         Examples
         -------
         >>> WOQLQuery().unique("https://base.url",["page","1"],"v:obj_id").execute(client)
-
         {'@type': 'api:WoqlResponse', 'api:status': 'api:success', 'api:variable_names': ['obj_id'], 'bindings': [{'obj_id': 'https://base.urlacd150a6885f609532931d89844070b1'}], 'deletes': 0, 'inserts': 0, 'transaction_retry_count': 0}
         """
         if prefix and prefix == "woql:args":
@@ -1933,7 +1932,6 @@ class WOQLQuery:
         Examples
         -------
         >>> WOQLQuery().idgen("https://base.url",["page","1"],"v:obj_id").execute(client)
-
         {'@type': 'api:WoqlResponse', 'api:status': 'api:success', 'api:variable_names': ['obj_id'], 'bindings': [{'obj_id': 'https://base.url_page_1'}], 'deletes': 0, 'inserts': 0, 'transaction_retry_count': 0}
         """
         if prefix and prefix == "woql:args":
@@ -1966,7 +1964,6 @@ class WOQLQuery:
         Examples
         -------
         >>> WOQLQuery().random_idgen("https://base.url",["page","1"],"v:obj_id").execute(client)
-
         {'@type': 'api:WoqlResponse', 'api:status': 'api:success', 'api:variable_names': ['obj_id'], 'bindings': [{'obj_id': 'http://base.url_page_1_rv1mfa59ekisdutnxx6zdt2fkockgah'}], 'deletes': 0, 'inserts': 0, 'transaction_retry_count': 0}
         """
         if prefix and prefix == "woql:args":
@@ -2340,7 +2337,7 @@ class WOQLQuery:
         self._cursor["woql:count"] = self._clean_object(countvar)
         return self._add_sub_query(query)
 
-    def cast(self, val, user_type, result):
+    def cast(self, val, user_type, result, literal_type=None):
         """Changes the type of va to type and saves the return in vb
 
         Parameters
@@ -2349,14 +2346,22 @@ class WOQLQuery:
             original variable
         user_type : str
             type to be changed
-        rsult: str
+        result: str
             save the return variable
+        literal_type: str, optional
+            literal type of`val`, can be used to treat `val` as a literal rather than an object or variable in the WOQL query.
+            If literal type is "owl:Thing" or "woql:node", `val` will be treated as object in the graph
 
         Returns
         -------
         WOQLQuery object
             query object that can be chained and/or execute
         """
+        if literal_type is not None:
+            if literal_type == "owl:Thing" or literal_type == "woql:node":
+                return self.cast(self.iri(val), user_type, result)
+            else:
+                return self.cast(self.literal(val, literal_type), user_type, result)
         if val and val == "woql:args":
             return ["woql:typecast_value", "woql:typecast_type", "woql:typecast_result"]
         if self._cursor.get("@type"):